requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==4.1.0
pyspark==3.4.3
psycopg2-binary==2.9.3
kafka-python==2.0.2